Start project
The goal of this project is to develop a library and starter bot
for the "Four in a Row" competition hosted by The AI Games.

This repository holds the code for this library, and will use
the GitFlow workflow to organize the branches. Also, semantic
versioning will be used to make it easy to differentiate between
releases and hotfixes.