Fix bug in bot
The starter bot is implemented without any real logic to it, and just
picks a random column to drop the disc in. It tried to get the number of
columns from the Match, but this class does not provide this
information. This led to a NoMethodError, which can easily be fixed by
requesting this information not from the Match, but from the
PlayingField.